arch/arm: unify some duplicate code to common layer
