+++ arch/arm/src/common/arm_releasepending.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * arch/arm/src/armv6-m/arm_unblocktask.c
+ * arch/arm/src/common/arm_releasepending.c
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
@@ -25,13 +25,12 @@
 #include <nuttx/config.h>
 
 #include <sched.h>
-#include <assert.h>
 #include <debug.h>
 #include <nuttx/arch.h>
 #include <nuttx/sched.h>
 
 #include "sched/sched.h"
-#include "clock/clock.h"
+#include "group/group.h"
 #include "arm_internal.h"
 
 /****************************************************************************
@@ -39,48 +38,34 @@
  ****************************************************************************/
 
 /****************************************************************************
- * Name: up_unblock_task
+ * Name: up_release_pending
  *
  * Description:
- *   A task is currently in an inactive task list but has been prepped to
- *   execute.  Move the TCB to the ready-to-run list, restore its context,
- *   and start execution.
- *
- * Input Parameters:
- *   tcb: Refers to the tcb to be unblocked.  This tcb is in one of the
- *     waiting tasks lists.  It must be moved to the ready-to-run list and,
- *     if it is the highest priority ready to run task, executed.
+ *   Release and ready-to-run tasks that have collected in the pending task
+ *   list.  This can call a context switch if a new task is placed at the
+ *   head of the ready to run list.
  *
  ****************************************************************************/
 
-void up_unblock_task(struct tcb_s *tcb)
+void up_release_pending(void)
 {
   struct tcb_s *rtcb = this_task();
 
-  /* Verify that the context switch can be performed */
+  sinfo("From TCB=%p\n", rtcb);
 
-  DEBUGASSERT((tcb->task_state >= FIRST_BLOCKED_STATE) &&
-              (tcb->task_state <= LAST_BLOCKED_STATE));
+  /* Merge the g_pendingtasks list into the ready-to-run task list */
 
-  /* Remove the task from the blocked task list */
-
-  nxsched_remove_blocked(tcb);
-
-  /* Add the task in the correct location in the prioritized
-   * ready-to-run task list
-   */
-
-  if (nxsched_add_readytorun(tcb))
+  if (nxsched_merge_pending())
     {
-      /* The currently active task has changed! We need to do
-       * a context switch to the new task.
+      /* The currently active task has changed!  We will need to
+       * switch contexts.
        */
 
       /* Update scheduler parameters */
 
       nxsched_suspend_scheduler(rtcb);
 
-      /* Are we in an interrupt handler? */
+      /* Are we operating in interrupt context? */
 
       if (CURRENT_REGS)
         {
@@ -88,7 +73,7 @@ void up_unblock_task(struct tcb_s *tcb)
            * Just copy the CURRENT_REGS into the OLD rtcb.
            */
 
-          arm_savestate(rtcb->xcp.regs);
+           arm_savestate(rtcb->xcp.regs);
 
           /* Restore the exception context of the rtcb at the (new) head
            * of the ready-to-run task list.
@@ -100,7 +85,9 @@ void up_unblock_task(struct tcb_s *tcb)
 
           nxsched_resume_scheduler(rtcb);
 
-          /* Then switch contexts */
+          /* Then switch contexts.  Any necessary address environment
+           * changes will be made when the interrupt returns.
+           */
 
           arm_restorestate(rtcb->xcp.regs);
         }
@@ -111,6 +98,15 @@ void up_unblock_task(struct tcb_s *tcb)
         {
           struct tcb_s *nexttcb = this_task();
 
+#ifdef CONFIG_ARCH_ADDRENV
+          /* Make sure that the address environment for the previously
+           * running task is closed down gracefully (data caches dump,
+           * MMU flushed) and set up the address environment for the new
+           * thread at the head of the ready-to-run list.
+           */
+
+          group_addrenv(nexttcb);
+#endif
           /* Update scheduler parameters */
 
           nxsched_resume_scheduler(nexttcb);
